backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["D.B. <derek.baugh1775@gmail.com>"]
readme = "README.md"
packages = [{include = "backend", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
quart = "^0.19.6"
websockets = "^12.0"

[tool.poetry.scripts]
start = "backend:run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
